ip_takeover
 python script to add or remove the VIP in the node function to his role
 master: add vip
 replica del vip

The script is written to be usable as a compiled Cython excutable.
modifying the addresses with the ip addr tool require root privilege.
It is not easy to run sudo on centos as sudo is reserved for uid > 1000
python script -> cython -> gcc -> exec file permit to have the exce file  SUID

On each node the vip must be defined in /usr/local/etc/patroni_floating_ip.conf:
one IPV6 per line (+ '%' + interface if the host has multiple interfaces)

patrony.yaml need to be modified to add in the postgresql section:
  callbacks:
    on_reload:
      /usr/local/sbin/ip_takeover
    on_restart:
      /usr/local/sbin/ip_takeover
    on_role_change:
      /usr/local/sbin/ip_takeover
    on_start:
      /usr/local/sbin/ip_takeover
    on_stop:
      /usr/local/sbin/ip_takeover

The VIP need to be added in all nftables rules as linux use the last
address added as the source address.